[Y26W2-349] refactor(web): 리스트 페이지 View 분리
